companyapi/requirements.txt
Django
djangorestframework
pytz
sqlparse
asgiref